requirements.txt
discord.py
psycopg
python-dotenv